+++ Source/Scene/Vector3DTilePrimitive.js
@@ -1055,8 +1055,7 @@ define([
 
         commandLength = commandsIgnoreShow.length;
         for (i = 0; i < commandLength; ++i) {
-            command = commandsIgnoreShow[i];
-            commandList.push(command);
+            commandList.push(commandsIgnoreShow[i]);
         }
     }