+++ Source/Shaders/BillboardCollectionFS.glsl
@@ -54,5 +54,5 @@ void main()
     gl_FragColor = color;
 #endif
 
-    czm_writeLogZ();
+    czm_writeLogDepth();
 }