+++ Source/Shaders/DepthPlaneVS.glsl
@@ -7,5 +7,5 @@ void main()
     positionEC = czm_modelView * position;
     gl_Position = czm_projection * positionEC;
 
-    czm_vertexLogZ();
+    czm_vertexLogDepth();
 }